Added support for WiFi poloarity and DHCP/USB switch.
currently the dts maps the switch as a button. This allows the user to use hotplug to
process switch movement events. The problem with this approach is that its impossible
to probe the switch on boot up for the the status.  So the user has to flip the switch
to cause an event.. the usb susbsitem can't be initialized on startup.

there are two solutions to this problem.
1. parse /sys/kernel/debug/gpio for the status of the switch on startup
or
2. not use the button at all and map the switch as a standard exported GPIO. If this is done
then the user has to write a monitor that polls the state of the switch and takes appropriate action.

1 is fine as long as the kernel is build with debugfs... it will fail if its not.

we need to decide which approach is best.

Signed-off-by: Luis Soltero <lsoltero@globalmarinenet.com>